Feat: moved constants into proper folder and updated url regex and config tests
