+++ packages/snap/src/constants/chainConfig.ts
